Create 053 - Trapping Rain Water - GFG
Given an array arr[] of N non-negative integers representing the height of blocks. If width of each block is 1, compute how much water can be trapped between the blocks during the rainy season. 
 
Example 1:
Input:
N = 6
arr[] = {3,0,0,2,0,4}
Output:
10
Explanation: 
block of height 0 is 3 units and above 
block of height 0 is 3 units. So, the 
block of height 2 is 1 units and above 
block of height 0 is 3 units. So, the 
total unit of water trapped is 10 units.

Example 2:
Input:
N = 4
arr[] = {7,4,0,9}
Output:
10
Explanation:
Water trapped by above 
block of height 4 is 3 units and above 
block of height 0 is 7 units. So, the 
total unit of water trapped is 10 units.